+++ src/main/java/com/example/sisrec/services/UsuarioService.java
@@ -2,6 +2,7 @@ package com.example.sisrec.services;
 
 import com.example.sisrec.models.UsuarioModel;
 import com.example.sisrec.repositories.UsuarioRepository;
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,24 +18,31 @@ public class UsuarioService {
         this.usuarioRepository = usuarioRepository;
     }
 
-//    public List<UsuarioModel> findAll() {
-//        return usuarioRepository.findAll();
-//    }
-
-
-    public Optional<UsuarioModel> buscarPorId(UUID id) {
-        return usuarioRepository.findById(id);
-    }
-
     public List<UsuarioModel> buscartodos() {
         return usuarioRepository.findAll();
     }
 
-    public void atualizar(UsuarioModel usuario) {
-        usuarioRepository.save(usuario);
+    public Optional<UsuarioModel> buscarUsuarioPorId(UUID id) {
+        return usuarioRepository.findById(id);
     }
 
-    public void deletar(UUID id) {
-        usuarioRepository.deleteById(id);
+    public Optional<UsuarioModel> atualizarUsuario(UUID id, UsuarioModel updatedUser) {
+        return usuarioRepository.findById(id).map(existingUser -> {
+            existingUser.update(updatedUser);
+            return usuarioRepository.save(existingUser);
+        });
+
+
     }
+
+    public Optional<UsuarioModel> deletarUsuario(UUID id) {
+        Optional<UsuarioModel> usuarioExistente = usuarioRepository.findById(id);
+        if (usuarioExistente.isPresent()) {
+            usuarioRepository.deleteById(id);
+            return usuarioExistente;
+        } else {
+            return Optional.empty();
+        }
+    }
+
 }
